requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.5.0
motor==3.3.1
pydantic>=2.6.4
python-multipart>=0.0.9
pdfplumber>=0.11.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0
